src/main/webapp/app/entities/category/update/category-update.component.ts
import { Component, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize, map } from 'rxjs/operators';

import { ICategory, Category } from '../category.model';
import { CategoryService } from '../service/category.service';
import { IFile } from 'app/entities/file/file.model';
import { FileService } from 'app/entities/file/service/file.service';

@Component({
  selector: 'jhi-category-update',
  templateUrl: './category-update.component.html',
})
export class CategoryUpdateComponent implements OnInit {
  isSaving = false;

  categoryFilesCollection: IFile[] = [];

  editForm = this.fb.group({
    id: [],
    name: [],
    description: [],
    thumbnail: [],
    active: [],
    categoryFile: [],
  });

  constructor(
    protected categoryService: CategoryService,
    protected fileService: FileService,
    protected activatedRoute: ActivatedRoute,
    protected fb: FormBuilder
  ) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ category }) => {
      this.updateForm(category);

      this.loadRelationshipsOptions();
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const category = this.createFromForm();
    if (category.id !== undefined) {
      this.subscribeToSaveResponse(this.categoryService.update(category));
    } else {
      this.subscribeToSaveResponse(this.categoryService.create(category));
    }
  }

  trackFileById(index: number, item: IFile): number {
    return item.id!;
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<ICategory>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(category: ICategory): void {
    this.editForm.patchValue({
      id: category.id,
      name: category.name,
      description: category.description,
      thumbnail: category.thumbnail,
      active: category.active,
      categoryFile: category.categoryFile,
    });

    this.categoryFilesCollection = this.fileService.addFileToCollectionIfMissing(this.categoryFilesCollection, category.categoryFile);
  }

  protected loadRelationshipsOptions(): void {
    this.fileService
      .query({ filter: 'category-is-null' })
      .pipe(map((res: HttpResponse<IFile[]>) => res.body ?? []))
      .pipe(map((files: IFile[]) => this.fileService.addFileToCollectionIfMissing(files, this.editForm.get('categoryFile')!.value)))
      .subscribe((files: IFile[]) => (this.categoryFilesCollection = files));
  }

  protected createFromForm(): ICategory {
    return {
      ...new Category(),
      id: this.editForm.get(['id'])!.value,
      name: this.editForm.get(['name'])!.value,
      description: this.editForm.get(['description'])!.value,
      thumbnail: this.editForm.get(['thumbnail'])!.value,
      active: this.editForm.get(['active'])!.value,
      categoryFile: this.editForm.get(['categoryFile'])!.value,
    };
  }
}
